refactor(cosmo_val): extract CosebisMixin (cosebis)
